# Specialize sign-related functions