This is part of the issue #185.
**Edit:**
This prohibits `copysign`/`flipsign` for `Normed`. They should be confusing when the signed `Normed` is supported in the future. Furthermore, if signed `Normed` is available, we should not do unsafe sign operations for unsigned `Normed`.