@kimikage 2020-06-28
### Benchmark
```julia
julia> versioninfo()
Julia Version 1.4.2
Commit 44fa15b150* (2020-05-23 18:35 UTC)
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-8.0.1 (ORCJIT, skylake)
```
#### before
```julia
julia> @btime copysign.(x, y) setup=(x=collect(rand(Q0f7,100,100)); y=collect(rand(Q0f7,100,100)));
  31.100 μs (1 allocation: 9.94 KiB)

julia> @btime flipsign.(x, y) setup=(x=collect(rand(Q0f7,100,100)); y=collect(rand(Q0f7,100,100)));
  30.900 μs (1 allocation: 9.94 KiB)
  
julia> @btime signbit.(x) setup=(x=collect(rand(Q0f7,100,100)));
  5.750 μs (3 allocations: 5.56 KiB)

julia> @btime signbit.(x) setup=(x=collect(rand(N0f8,100,100)));
  2.744 μs (3 allocations: 5.56 KiB)
  
julia> @btime abs.(x) setup=(x=collect(rand(N0f8,100,100)));
  2.962 μs (1 allocation: 9.94 KiB)
```

#### after
```julia
julia> @btime copysign.(x, y) setup=(x=collect(rand(Q0f7,100,100)); y=collect(rand(Q0f7,100,100)));
  7.500 μs (1 allocation: 9.94 KiB)
  
julia> @btime flipsign.(x, y) setup=(x=collect(rand(Q0f7,100,100)); y=collect(rand(Q0f7,100,100)));
  5.350 μs (1 allocation: 9.94 KiB)
  
julia> @btime signbit.(x) setup=(x=collect(rand(Q0f7,100,100)));
  5.316 μs (3 allocations: 5.56 KiB)

julia> @btime signbit.(x) setup=(x=collect(rand(N0f8,100,100)));
  1.700 μs (3 allocations: 5.56 KiB)

julia> @btime abs.(x) setup=(x=collect(rand(N0f8,100,100)));
  2.962 μs (1 allocation: 9.94 KiB)
```